Adds exception handling around the scan for loop.
This allows the NIDAQ tasks to be closed gracefully after an exception.
Doing this frees the resources to be used by another application or
caller.

This is especially useful when using the CWODMR class within a
Jupyter notebook as it allows the user to stop execution of the
kernel (which raises a KeyboardInterupt error) without any side-effects.

Data acquired up to that point should be returned after the exception
and the CWODMR.run function can be called immediately afterward without
restarting the kernel or other.