Simplifies scanning object to generic CounterAndScanner
Allows to be instantiated with an implementation of a RateCounterBase
class. RandomRateCounter and NiDaqDigitalInputRateCounter are subclasses.

Also, CounterAndScanner must be implemented with a stage_controller object that
is a subclass of nipiezojenapy.BaseControl. (This is poor choice,
as we should probably define an interface for a scanner and then
an implementation that supports nipiezojenapy. This would allow for
different scanners to be used in the future. See Issue #79)