Updates conversion of data lists to numpy arrays.
After data acquisition is complete, data is stored in a
list where each element of the list is a list of length
two. The first element is a float and the second element
is a numpy array. New versions of numpy prohibit
automatic conversion of this kind of list to a numpy array
without explicitly stating the data type of the array
as an 'object'. This change is necessary to use versions
of numpy > 1.23.