Adds hyper-data acquisition function to piezoscanner.
A number of changes have been made in this commit to support
the addition of a hyper-data acquisition callback function.

This function allows an external user to define a function
to be called at each position of the scan. This function
may also be executed in a separate thread if the 'in_parallel'
attribute of the funciton is set.

In support of this function, a number of other smaller changes
were made.

A run_scan function was added in order to facilitate use
of CounterAndScanner objects. A callback function for that
is also provided so the GUI application pizeoscan.py may
update it's display upon completion of a raster line.

Unnecessary checks for stage_controller != None were removed
since this object does not support instantiation without
a stage_controller. (Although, future updates should provide
proper type hinting.)

The position of the scan is no longer tracked by the CounterAndScan
object, but instead relies upon the last_write_values attribute
of the stage_controller object.