Changes internal naming.
A misconception about the hardware led to incorrect variable
names and documentation. Initially, the piezo actuator
that controls the position of the laser light impinging upon
a sample was called a "stage" controller due to a misunderstanding
of the hardware setup. This labeling perpetuated until this change.
All usage of "stage" control has been replaced either with
"position" control or actuator, where appropriate.
Variable naming may still feel a bit off in places. Please fix
if you have better ways of describing functionality.