qt3scan supports more hardware configurations.
Previously, the code was restricted such that you could
only load a yaml file that configured the currently selected
hardware controller in qt3scan. For example, you selected
the random data generator controller, you could only load
a yaml file that provided configuration for QT3ScanDummyPositionController
and QT3ScanRandomDataController.

Now one can use different position controllers and daq
controllers for the same application controler. That is, with
the 'random data generator' controller selected, one can load
a yaml file that defines using a QT3ScanNIDAQPositionController
as the position controller.

Additionally, the 'NIDAQ Edge Counter' application controller
could be used with a QT3ScanDummyPositionController, for example.

The benefit of this is to support controllers that may be developed
in the future which can be directly swapped with currently defined
controllers. This is allowed so long as the application controller,
as selected by the pull down menu, can use the position or daq
controller in an expected way. That is, that controller adheres
to the proper interface.

It's still highly encouraged that if you develop your own daq,
or position controller, that you add it to qt3-utils so that others
may share from your tools.