WIP-Removes clock_rate property from qt3scan interface.
This property is not necessary for functionality of qt3scan.
However, this doesn't mean a similar property isn't necessary
for some daq components (NiDaqDigitalInputRateCounter).

This has been tested on a local machine with random data simulator
but must be tested on actual hardware before merging -- modify
this commit message once testing is completed.