wip - fixes selection of correct hyperspectral image when right-clicking on the application.
still not perfect as the application controller has to
adjust by step_size/2 because qt3scan has (correctly)
adjusted the scan image by step_size/2.

Need a solution: either qt3scan "fixes" the xy value
of the mouse click before sending to the callback function,
or passes the step_size/2 value as a second arg to the
callback funciton, OR the 2D image does NOT make any
adjustment at all in the first place. All three solutions
have pros/cons.

One thing we have to consider is the position controller!
When a user clicks on the 2D image as says "go to position",
it will go to that exact x,y value. However, if the image
is not property centered over the axis, then the position
of the conroller will be shifted slightly away from the
intended location. Thus, the step_size/2 fixes this problem
and allows the user to actually move to the desired location.